got write batching to pass tests and added file/buffer position recording

+++ disk/src/com/treode/disk/edit/PageDispatcher.scala
@@ -28,9 +28,9 @@ import com.treode.async.{Async, Callback, Fiber, Scheduler }, Async.async
 private class PageDispatcher(private var counter: Long)
  (implicit
   scheduler: Scheduler
-) extends Dispatcher [(String, Callback[(Long, Int)])] 
+) extends Dispatcher [(String, Callback[(Long, Long)])] 
 {
-  def write (data: String) : Async [(Long, Int)] = {
+  def write (data: String) : Async [(Long, Long)] = {
     async { cb =>
     val batch = (data, cb)
     send(batch)
